Hacky fix for DacapoClassLoader
This commit fixes DacapoClassLoader for some benchmarks (such as
luindex) in a hacky way. We create a chain of DexClassLoaders and search
for a matching dex file in reverse order as otherwise the standard class
loader is not dex file aware.

This commit also tried to build Lucene for Android, however the Android
stdlib is missing many standard Java 11 classes such as
java.lang.Runtime.Version and java.lang.Module. Hence, while we can
compile Lucene for Android, Lucene will not actually run as it will
error with NoSuchMethod or ClassNotFound exceptions.